Explicitly check whether we actually have a rate before caching it
This would especially happen around midnight UTC, where the api we use
would already return a json object for the new day. But fill it without
any data. And as it turns out, if your datatype is float64 and you retrieve
it from a map. It will just be 0 if it's not there.. So we now explicitly check
whether it's in the map before we cache it.